+++ R/gads_utils.R
@@ -1,7 +1,7 @@
 # cache
 gads_get_fields_cached <- memoise(gads_get_fields)
 
-# make query
+# make gaul query
 gads_make_query <- function(
   resource              = 'campaign',
   fields                = c('campaign.id',
@@ -82,5 +82,60 @@ gads_make_query <- function(
        )
 
   return(gaql_query)
+
+}
+
+
+# make request
+gads_make_request <- function(
+    http_method       = "POST",
+    body              = NULL,
+    api_version       = options("gads.api.version"),
+    api_method        = "googleAds:searchStream",
+    customer_id       = getOption('gads.customer.id'),
+    login_customer_id = getOption('gads.login.customer.id'),
+    verbose           = TRUE
+  ) {
+
+  # send query
+  out <- request_build(
+    method   = http_method,
+    body     = body,
+    path     = str_glue('{api_version}/{api_method}'),
+    token    = gads_token(),
+    base_url = getOption('gads.base.url')
+  )
+
+  # send request
+  ans <- request_retry(
+    out,
+    encode = 'json',
+    add_headers(
+      `developer-token`= gads_developer_token(),
+      `login-customer-id` = login_customer_id
+      )
+  )
+
+  # --------------
+  # get answer
+  if (verbose) cli_alert_info('Get answer query')
+  out <- response_process(ans, error_message = gads_check_errors2)
+
+  # requests_ids
+  if ( !is.null(ans$headers$`request-id`) ) {
+
+    rq_ids <- unique(ans$headers$`request-id`)
+    rgoogleads$last_request_id <- rq_ids
+    if (verbose) cli_alert_info(c("Your request ids: ", rq_ids))
+
+  } else {
+
+    rgoogleads$last_request_id <- unique(sapply(out, function(x) x$requestId))
+    rq_ids <- str_c(rgoogleads$last_request_id, collapse = ', ')
+    if (verbose) cli_alert_info(c("Your request ids: ", rq_ids))
+
+  }
+
+  return(out)
 
 }